Added files that are in the raspberry pi pico.
This includes all the libraries necessary to run the code too. Such as the font file needed to display text, the actual library files required for the OV7670 module to work. To switch between the models, you just need to change svm_min.py file in this directory. This current model is for classifying handwritten digits.